requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
blis==0.7.8
boto3==1.24.57
botocore==1.27.57
cachetools==4.2.4
catalogue==2.0.8
charset-normalizer==2.1.0
click==8.1.3
cymem==2.0.6
dataclasses==0.6
datasets==2.4.0
dill==0.3.5.1
docker-pycreds==0.4.0
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
keras==2.9.0
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.7
nltk==3.7
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
pathy==0.6.2
preshed==3.0.6
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-lightning==0.9.0
pytz==2022.1
PyYAML==5.4.1
regex==2022.7.24
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.0
sentencepiece==0.1.91
sentry-sdk==1.9.4
seqeval==0.0.12
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
tensorboard==2.2.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.21.0
truecase==0.0.14
typer==0.4.2
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.11
wandb==0.13.1
wasabi==0.10.1
Werkzeug==2.2.0
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.1